refactor: BuyMeACoffeButton.vue to BuyMeACoffeeButton.vue
